I created a new Python file and put shared functions in there, so they are shared with all the unit test files that test file logging. 2. Removed some blank line between decorator and the function it is decorating. 3. Removed a comment.
